Move About.js to Main folder with related file adjustments
